requirements.txt
asgiref==3.7.2
Django==4.2.2
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
Pillow==9.5.0
PyJWT==2.7.0
pytz==2023.3
sqlparse==0.4.4
typing_extensions==4.6.3
gunicorn==20.1.0
mysqlclient==2.0.3
